Introduce the concept of "capabilities"
This commit introduces a DSL to declare "capabilities" for modules. Capabilities can be used to declare
that different modules, published at different coordinates, actually provide the same capabilities. This
is typically the case when an API is implemented in different ways (for example, the Slf4j bindings
implement the Slf4j API). It is an error to have multiple modules providing the same capabilities in the
same dependency graph: only one of them should be selected.

This commit adds the ability to declare capabilities, explaining which modules provide the capability, and
provide a way to select the module which will provide that capability. It does NOT make it an error to have
different modules providing the same capability on the graph yet (if a preference is not set).

This solves different use cases:

1. module relocation (`asm:asm` is now found at `org.ow2.asm`)
2. multiple variants of the same module are published at different coordinates (`cglib` vs `cglib-nodep`)
3. multiple implementations of the same API are seen in the graph (`slf4j-simple`, `slf4j-over-log4j`)
4. compatiblity module mimics a different API (log4j2 bridge for log4j1)