Report error whenever two modules in the graph disagree on a preference
This commit makes it an error case to have two modules provide a preference for a capability, if that preference
is not the same. Currently we don't know how to solve such a problem, in particular because preferences can be
discovered as the graph grows, and are hard to rollback.